Changes Table + New Account Styles
New Tables (change_data and attribute_change) at the SQL to storage the changes executed by an admin
New account.css Styles , make it more modern

+++ public/cart/removeProduct.php
@@ -2,37 +2,44 @@
     include "../../databaseConnection.php";
     include "../generalPHP.php";
 
-    $getAllProducts = $mysqli->query("SELECT nameProduct FROM product_version");
-    $allowedProducts = [];
-    while($allProducts = $getAllProducts->fetch_assoc()){
-        $allowedProducts[] = $allProducts["nameProduct"];
-    }
+    if(isset($_GET["name"])){
+        $getAllProducts = $mysqli->query("SELECT nameProduct FROM product_version");
+        $allowedProducts = [];
+        while($allProducts = $getAllProducts->fetch_assoc()){
+            $allowedProducts[] = $allProducts["nameProduct"];
+        }
 
-    $getAllProducts->close();
+        $getAllProducts->close();
 
-    if(in_array($_GET["name"], $allowedProducts)){
-        $removeProd = $mysqli->prepare("
-            DELETE FROM product_order 
-            WHERE idProduct = (
-                SELECT idVersion 
-                FROM product_version
-                WHERE nameProduct = ?
-            )
-            LIMIT 1;
-        ");
+        if(in_array($_GET["name"], $allowedProducts)){
+            $removeProd = $mysqli->prepare("
+                DELETE FROM product_order 
+                WHERE idProduct = (
+                    SELECT idVersion 
+                    FROM product_version
+                    WHERE nameProduct = ?
+                )
+                LIMIT 1;
+            ");
 
-        $removeProd->bind_param("s", $_GET["name"]);
+            $removeProd->bind_param("s", $_GET["name"]);
 
-        if($removeProd->execute()){
-            $removeProd->close();
-            header("location: cart.php");
-            exit();
+            if($removeProd->execute()){
+                $removeProd->close();
+                header("location: cart.php");
+                exit();
+            }else{
+                $removeProd->close();
+                header("../errorPage.php");
+                exit();
+            }
         }else{
-            $removeProd->close();
             header("../errorPage.php");
             exit();
         }
     }else{
-        header("../errorPage.php");
-        exit();
+        header("location: cart.php");
+        exit;
     }
+
+
